refactor(generator): 하위 패키지가 상위 패키지를 참조하지 않도록 구조를 바꾼다.
