+++ src/main/java/racingcar/config/AppConfig.java
@@ -3,8 +3,8 @@ package racingcar.config;
 import racingcar.controller.RacingController;
 import racingcar.domain.car.move.DefaultMovingStrategy;
 import racingcar.domain.car.move.MovingStrategy;
-import racingcar.domain.generator.NumberGenerator;
-import racingcar.domain.generator.RandomNumberGenerator;
+import racingcar.domain.car.move.generator.NumberGenerator;
+import racingcar.domain.car.move.generator.RandomNumberGenerator;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 import racingcar.view.console.ConsoleInputView;
