+++ src/main/java/racingcar/domain/car/move/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.generator;
+package racingcar.domain.car.move.generator;
 
 public interface NumberGenerator {
     int generate();
